BUG Fix hitmiss for non-boolean inputs
This was just a typo in the code.

Reported by Jean Patrick Pommier on pythonvision.

+++ mahotas/morph.py
@@ -157,7 +157,7 @@ def hitmiss(input, Bc, output=None):
             else:
                 Bc = Bc.astype(np.uint8)
         else:
-            Bc = Bc.astype(np.dtype)
+            Bc = Bc.astype(input.dtype)
     if output is None:
         output = np.empty_like(input)
     else:
